ai-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
crewai>=0.28.0
langchain-groq>=0.1.0
pydantic>=2.0.0
